Srikrishnudu_nagandla/Hackathon/multi_model_ai_design_analysis/requirements.txt
streamlit==1.29.0
faiss-cpu==1.7.4
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
langgraph==0.0.20
langchain==0.1.0
langchain-core==0.1.10
torch==2.1.0
torchvision==0.16.0
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1